# [BBR] perf: optimize model name extraction with selective JSON unmarshaling using struct tags